Fix FIN/opcode in wsframe_write (issue #1)
This is a quick bug fix to a bug reported by Mikwind.
Now satisfies the spec for closing the report. However, the
fix is not complete. See below.

I guess my intention was to do this sort of a thing as a way of
writing a TODO-note to myself, as this may document I wanted to set
FIN to 1 and opcode to 1, while RSV1-3 in the middle should be 0.
However, as I had it there, it results in 00000001 in binary, which
does not result in correct preamble as intended.

Looking at the example usage in the ifdef'd TEST block, FIN should
be 1 because only one packet is written. It is obviously unfinished
code and the behavior should be modified so that FINs are written
properly. The correct hexadecimal number for 10000001 is 0x81 for
setting FIN=1 and opcode=1 and interestingly it was used correctly
in the test byte sequences in the TEST block.

Chances are, the code in the wsframe_write was partially phased
out, since good logic seems to be already happening in wsframe_add_bytes
which does the correct thing with FIN and opcode. What happens in
wsframe_add_bytes should be reflected in the wsframe_write that
obviously should be re-written since the current one is only
half-finished that works only for getting some reactions when
testing, but is not useful for anything real. Should be fairly
trivial to wrap it up for good, though.